+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### v0.5.0
+- :bulb: Added status field for contacts concepts
 - :hammer: Fixed dockerfile assets precompile
 - :hammer: Fixed minor bugs